% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Mult_Test_Stat.R
\name{Compute_Mult_Test_Stat}
\alias{Compute_Mult_Test_Stat}
\title{Quantify indirect effect using inverse odds weighting}
\usage{
Compute_Mult_Test_Stat(exposure, outcome, mediator)
}
\arguments{
\item{exposure}{vector of exposure variable}

\item{outcome}{vector of outcome variable}

\item{mediator}{matrix of mediator variables}
}
\value{
The indirect effect of multiple mediators
}
\description{
This function performs mediation analysis using inverse
odds weighting to compute the indirect effect of multiple mediators.
}
