% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_Mult_Test_Stat_adj.R
\name{Compute_Mult_Test_Stat_adj}
\alias{Compute_Mult_Test_Stat_adj}
\title{Quantify indirect effect using inverse odds weighting adjusted for confounders}
\usage{
Compute_Mult_Test_Stat_adj(exposure, outcome, confounder, mediator)
}
\arguments{
\item{exposure}{vector of exposure variable}

\item{outcome}{vector of outcome variable}

\item{confounder}{matrix of covariates}

\item{mediator}{matrix of mediator variables}
}
\value{
The indirect effect of multiple mediators
}
\description{
This function performs mediation analysis using inverse odds weighting to
compute the indirect effect of multiple mediators with the presence of confounders.
}
