% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMMIOW.R
\name{Perform_HTMMIOW}
\alias{Perform_HTMMIOW}
\title{Perform HT-MMIOW}
\usage{
Perform_HTMMIOW(
  exposure,
  outcome,
  confounder = NULL,
  mediator,
  umapcomponents = 2,
  B = 2000
)
}
\arguments{
\item{exposure}{vector of exposure variable}

\item{outcome}{vector of outcome variable}

\item{confounder}{matrix of relavant covariates}

\item{mediator}{matrix of integers that represent abundance}

\item{umapcomponents}{number of components for UMAP to reduce to}

\item{B}{number of replicates for hypothesis test}
}
\value{
A list of observed test statistic, p-value, and test statistics under the null hypothesis
}
\description{
This function performs HT-MMIOW on user-provided exposure, outcome,
confounder, and mediator data.
}
\examples{
Perform_HTMMIOW(exposure = sim_data$sim_exposure,
outcome = sim_data$sim_outcome,
mediator = sim_data$sim_microbiome,
umapcomponents = 2)
}
