% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated exposure, outcome, and microbiome data}
\format{
A list of exposure, outcome, and microbiome data for 300 subjects and 400 microbes.
}
\usage{
sim_data
}
\description{
A dataset containing dichotomous exposure, dichotomous outcome,
and microbiome abudance data. Exposure and microbiome data
simulated using biobakery's SparseDOSSA2.
}
\references{
Ma et al. (2021)  PLoS Comput Biol 17(9): e1008913
(\href{https://pubmed.ncbi.nlm.nih.gov/34516542/}{PubMed})
}
\keyword{datasets}
