% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTMMIOW.R
\name{Perform_HTMMIOW}
\alias{Perform_HTMMIOW}
\title{Perform HT-MMIOW}
\usage{
Perform_HTMMIOW(
  exposure,
  outcome,
  confounder = NULL,
  mediator,
  umapcomponents = 2
)
}
\arguments{
\item{exposure}{vector of exposure variable}

\item{outcome}{vector of outcome variable}

\item{confounder}{matrix of relavant covariates}

\item{mediator}{matrix of mediator variables}

\item{umapcomponents}{number of components for UMAP to reduce to}
}
\value{
A list of observed test statistic, p-value, and test statistics under the null hypothesis
}
\description{
This function performs HT-MMIOW on user-provided exposure, outcome,
confounder, and mediator data.
}
\examples{
Perform_HTMMIOW(exposure = sim_data$sim_exposure,
outcome = sim_data$sim_outcome,
mediator = sim_data$sim_microbiome,
umapcomponents = 2)
}
